refactor: reorganize Matching-related packages from `entity` to `model` and update imports
