+++ src/main/java/com/yeoljeong/tripmate/matching/infrastructure/persistence/jpa/SpringDataMatchingRepository.java
@@ -1,6 +1,6 @@
 package com.yeoljeong.tripmate.matching.infrastructure.persistence.jpa;
 
-import com.yeoljeong.tripmate.matching.domain.entity.Matching;
+import com.yeoljeong.tripmate.matching.domain.model.Matching;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
